+++ src/routes/index.ts
@@ -1,9 +1,15 @@
 import { Router } from 'express';
-import v1Routes from './v1';
+
+import routers from '@/config/routers';
+import type { APIVersion } from '@/shared/types/route.type';
 
 const router: Router = Router();
 
-// Mount versioned routes on the main router
-router.use('/v1', v1Routes);
+// Register all routes
+(Object.keys(routers) as APIVersion[]).forEach((version) => {
+  routers[version].forEach((r) => {
+    router.use(`/api/${version}${r.path}`, r.route);
+  });
+});
 
 export default router;