+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-10-07
+
+### Fixed
+
+- Ensured express-mongo-sanitize can safely modify req.query.
+
+### Changes
+
+- Refactored route structure to separate definitions by version (`v1`) in `routes/config/routers.ts`.
+- Added APIVersion, RouteItem, and Routers types.
+- Automated route registration in `routes/index.ts`.
+
+---
+
 ## [1.0.1] - 2025-10-06
 
 ### Changed